requirements.txt
# Django
Django==4.2
djangorestframework==3.15.1

pillow==10.3.0
djangorestframework-simplejwt==5.3.1


